Discrete Mathematics Shell
This shell requires the user to enter two arrays in order to function
and in some options there must be a value for the indexed search. There
are many options here and updates will be done when they become
available. It's strongly encureged to get involved in the project as
there arent many like these out there.